requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.124.0
frozenlist==1.8.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.7.0
natsort==8.4.0
notifiers==1.3.6
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rpds-py==0.30.0
setuptools==80.9.0
six==1.17.0
starlette==0.50.0
treelib==1.8.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
wheel==0.45.1
yarl==1.22.0
zipp==3.23.0
